Fix overflow in delayMicrosecondsInterruptible function.
The left shift of by 2 of the variable us overflows for any
value > 0x3fff.

Remove shift and increase loop to 16 clocks per cycle.

This fix still has the problem that the delay depends on the
system clock being 16 mhz. This should be restructured to use
a frequency independent approach.