More simulator work.
Generalize the simulator pin definitions to divorce us more from
the specific chip being used. The pins defined in config.h are
ignored and replaced with stand-in enums referencing control
arrays in the simulator. Heater pins are replaced with standins,
too, but they overlap the other pins. Need to check if that's ok.