Replace sim_timer with AVR-simulator timer
The original sim_timer and sim_clock provided direct replacements
for timer/clock.c in the main code.  But when the main code changed,
simcode did not.  The main clock.c was dropped and merged into timer.c.
Also, the timer.c now has movement calculation code in it in some
cases (ACCELERATION_TEMPORAL) and it would be wrong to teach the
simulator to do the same thing.  Instead, teach the simulator to
emulate the AVR Timer1 functionality, reacting to values written to
OCR1A and OCR1B timer comparison registers.

Whenever OCR1A/B are changed, the sim_setTimer function needs to be
called.  It is called automatically after a timer event, so changes
within the timer ISRs do not need to bother with this.

A C++ class could make this requirement go away by noticing the
assignment.  On the other hand, a chip-agnostic timer.c would help
make the main code more portable.  The latter cleanup is probably
better for us in the long run.