Do not process and acknowledge empty lines
Some hosts sends lines that are not commands, some ends lines with
CRLF. Teacup send ACK on each CR or LF char even if line was empty.
Working with (for example) Repetier Host it causes problems because
host count OKs form firmware and calulates how many bytes it could
transfer. When one line is acknowledged twice it will desynchronize
communication and will cause errors.
This code correct this "bug". It's only a workaround, but it works
OK.
TODO Decide if to send OK inside process_gcode_command and return
value to gcode_parse.