Configtool: improve search for avr-gcc & Co.
This should allow to build on any OS with Arduino IDE installed
and/or with avr-gcc & Co. available trough the normal shell.

Instead of looking at the current platform it's looked at wether
an Arduino IDE path is set. If yes, tools inside there are used.
If no such path is set, just the basename is used to let the OS
search for it ($PATH, etc.).

This is related to issue #152.