serial-arm.c: allow sending arbitrarily long messages.
On ARM we use only the 16 byte hardware buffer for sending and
receiving over the serial line, which is often too short for
debugging messages. This implementation works fine and still
neither blocks nor introduces delays for short messages.

Costs 72 bytes binary size, mostly because it's the first usage
of delay_us():
    SIZES          ARM...     lpc1114
    FLASH  :  1656 bytes           6%
    RAM    :   136 bytes           4%
    EEPROM :     0 bytes           0%