Tools: add git-regtest.
This is a tool to run regression tests over a whole series of
commits. Typical usage:

  git checkout <topic branch>
  tools/git-regtest experimental

This runs 'make regressiontest' for every single commit between
experimental and HEAD of the topic branch. Very convenient to make
sure not a single commit introduces regressions.

If there are regressions, the tools stops at the commit bringing
in the regression, so you immediately know where to look.