Step timer: reset timer after pauses instead of doing a guess.
We know already wether we start from a pause or not, so let's
take advantage of this knowledge instead of checking for
plausibility of a timer delay at interrupt time.

Costs just 8 bytes binary size:

    SIZES          ARM...     lpc1114
    FLASH  :  7764 bytes          24%
    RAM    :   960 bytes          24%
    EEPROM :     0 bytes           0%

Due to the less code at interrupt time, maximum step rate was
raised from 127.9 kHz to 130.6 kHz.