ARM: allow non-PWM pins as heater output.
Just like with AVR, they're simply set as a GPIO output and
turned on and off as needed.

Bed heater and temp sensor not yet re-enabled, because Gen7-ARM
has a driver MOSFET for the bed, which needs an inverted signal.
It can be enabled for testing; M106 P1 S2 turns it on, M106 P1 S20
turns it off. Not the way it should work.