MOVEBUFFER_SIZE really does not have to be power of 2
A comment in dda_queue.h says that MOVEBUFFER_SIZE no longer needs to be
a power of 2 in size.  The comment is from a commit in 2011, but only
queue_full seems to be modified to make this true.  Other places in the
code still assume that MOVEBUFFER_SIZE is always a power of 2, wrapping
it with "& (MOVEBUFFER_SIZE-1)".

Add a MB_NEXT(x) macro which can be used to definitively find the next
slot in the queue without using any boolean math.  Replace all the
queue-position functions with new code that uses this MB_NEXT function
instead.

Also change the queue_full function to use this simpler method instead
of the complicated multi-step confusion which it did.