Reduce lookahead "error" to "notice"
The marlin firmware reportedly reports "Error" instead of "!!", indicating
a machine failure which is followed by a full power-down.  The Octoprint
GCode-sender assumes a reported Error means the print has failed and
the machine turned off.

In Teacup we report an "Error" when lookahead was too slow to join
movements, but this is interpreted as an emergency-stop by Octoprint who
then stops the job and leaves the printer idle with all the heaters running.

Change this "Error" to a "Notice" to avoid this problem.  Add a comment
prefix while we're at it to fit the de facto standard better.

See http://reprap.org/wiki/G-code