Optimize ADC temperature search
Use a binary search to find our target temperate in fewer
comparisons.  The search algorithm is ostensibly slower because
it involves a division, but it's a div-by-two so should be optimized
into a simple bit-shift.  Fewer comparisons involves fewer pgm_read_words
and should be faster overall, but the gain is arguably tiny.