Optimize ADC temperature search
Use a binary search to find our target temperate in fewer
comparisons.  The search algorithm is ostensibly slower because
it involves a division, but it's a div-by-two so should be
optimized into a simple bit-shift.  Fewer comparisons involves
fewer pgm_read_words and should be faster overall, but the gain
is arguably tiny.

Note by Traumflug:

According to @Wurstnase's performance measurements, gain is
actually pretty huge. Temp conversions 0..1024 before this
and the previous few commits:

  minimum: 1011.84 clock cycles.
  maximum: 1993.92 clock cycles.
  average: 1768.85 clock cycles.

Now:

  minimum: 437.72 clock cycles.
  maximum: 494.76 clock cycles.
  average: 470.278 clock cycles.

That's a speedup by factor 3.7 on average!