Made heaters update every 250ms for all sensor types
Heater update procedure heater_tick() was being called in varying intervals,
depending on heater's temperature sensor type and its sampling frequency.
However, PID code assumes it's being called in regular intervals of 250ms. Here,
it was being called in 10ms, 100ms, or 250ms intervals, depending on sensor type.
This caused calculations to be wrong and basically made PID controller work the same
as bang-bang, only perform more computation.

In this commit, I've added a separate timer/counter for each temperature sensor.
The counter is responsible for measuring time between heater_tick() calls. This
allows to update the heater independently of the sensor and in regular, fixed
intervals of 250ms.