Properly handle temp sensor initialization
The existing code eschews "break" apparently thinking that it will
exit the loop prematurely.  In fact this will only exit the switch
correctly.  The loop will run as normal.

Maybe it's not a big deal, though, since the sensors get re-inited
each time the type of sensor is encountered in the loop.  Presumably
it's ok for each of these sensors to get inited multiple times. If
that's always the case, maybe there's no problem here. But with the
current code sensor types might be initialized even when there is no
sensor defined for that type if the TEMP_* boolean is #defined.

For example, if I have this in my config:

    #define TEMP_MAX6675
    #define TEMP_MCP3008
    //#define TEMP_INTERCOM

    DEFINE_TEMP_SENSOR(extruder, TT_MAX6675, AIO5,  0)
    DEFINE_TEMP_SENSOR(bed,      TT_MAX6675, AIO7,  0)

Then the MAX6675 initialization code will be run twice and the MCP3008
code will, too, even though no temp sensor is defined for that type.

This change does not address the "init runs twice" problem.  It only
addresses the "initialization of unused types" problem.