temp.c: cleanup temp_sensor_tick() and next_read_time sanity.
Integrate the next_read_time countdown into the loop as is common.

Check for start_adc() in the same loop -- before decrementing the
timer -- and call it when needed on the tick before we need the
results.

One concern I have still is that start_adc() may be called twice
within a few microseconds if two probes need to be read. I expect
it should only be called once, but I am not readily familiar with
the AVR ADC conversion protocol.