temp.c: refactor temperature reads into functions.
Every type of temp sensor has its own special way to be read
and the once-simple loop has grown complex.  Restore some sanity
by isolating the code for each sensor into its own inline
function.

In the process I noticed temp_flags is only ever set (never read)
and is used only in MAX6675.  I don't understand what it was used
for, so for now, let's comment it out and revisit this later in
this series.