Simplify usb serial_writechar for clarity
We have already moved the "can we send" check into the front of the
loop, but we have another atomic region before that dedicated only to
deciding whether we should quit early because we timed out on the last
attempt and we find we are still not ready; or we find that we are
ready and so we clear the previously-timed-out flag.

Simplify this by clearing the flag inside the try-to-send loop coming
right after this code and then doing the check for previously-timed-out
early return after the attempt-to-send.  When we reversed the logic inside
the loop from "wait if not-ready" to "send when ready" this code path to
the early return becomes simple and just as fast as it was before.  It
does mean we also check it needlessly on every subsequent iteration of
this loop, but we are timed-waiting for a timeout at that point so it
is of little consequence.

It should be safe to move the "timeout=" calculation outside the atomic
region since it is clear we check this value outside the atomic region,
also.  This is the only thing left in the first atomic region in this
function and that seems unnecessary now.  So go ahead and drop the atomic
markers there.