dda_clock: Modify dda->n only in atomic code
dda_clock might be interrupted by dda_step, and dda_step might use or
modify variables also being used in dda_clock.  In particular dda->c
is modified in both functions but it is done atomically in dda_clock
to prevent dda_step from interrupting during the write.  But dda->n
is also modified in both places and it is not protected in dda_clock.

Move updates to dda->n to the atomic section along with dda->c.