thermistortable: allow negative slope
We assumed the slope of the thermistor linear approximation tables is
negative, but this prevents us from working with positively correlated
thermistor table values.

Add support for positively correlated temperature probes by allowing
the slope to be positive or negative for temperatute graphs. Change
the table generation in configtool to support this; change the default
sign of the slope to positive meaning previously generated values now
need to be regenerated to be correctly negative.

The new signed type necessitates a structure to hold the data rather than
a simple array.

The compiler zero-fills the slope when we omit it, even in a structure, so
use the zero to detect old-style tables instead of the size of the
defined structure.  Unfortunately this means the code size is 400 bytes
larger so long as we still support slope-less tables, regardless of
whether the user has included one.