WIP: acceleration work, mostly bogus
After lots of debugging for accuracy, I've decided this is a bit of a waste
so long as we still use dda_clock to do all of our math. The reason is
that the number of steps which run at some velocity are arbitrary based on
our TICK_TIME interval.  Acceleration from zero works fine, but deceleration
from speed does not. We decelerate fine but we run out of steps too soon
before we get to our target velocity of zero.

The trick to "do all the math in dda_clock; it's fast enough" works for
the current code which calculates velocity(step_number). But it does not
work so well for this change which tries to calculate velocity(time). This
is because the step_number is not consistent with the time intervals.

We could force it to align only by changing the length of our time_tick, or
if we could find a clever way to adjust our deceleration constant on the
way down so our "extra" steps do properly align with our final V=0. But
this math seems more complex than we want and difficult to adapt to
different configurations automatically.

You can see the error with this:

    grep elapsed datalog.out |
      sed -e 's/.*elapsed=\([0-9]*\)  *vel=\([0-9]*\) .*/\1 \2 /' > vplot
    gnuplot -p -e "plot 'vplot' using 1:2 with steps"

In this graph we should see velocity increase from zero and return to
almost-zero before starting the next move.  Instead we see velocity remains
several intervals above zero at the end of each move.  It is close, but
I expect it is inconsistent and a matter of hand-tuned luck at this point.

Maybe I'm overthinking that.  For now I want to just put this aside and
consider other avenues.

One option is to use a linear approximation table to get us around the
inv-sqrt curve and then further apply this to dda->c at each step.  This
would get us back in the business of calculating the next step duration
in dda_step, though.

Another option, maybe, is to have dda->next_phase_c which is used after
some number of steps is passed. This would let us ensure we have the same
number of steps during deceleration as we had during acceleration, and we
should then arrive at the final step exactly at zero. This begins to feel
quite hacky, but it also sounds workable.