Synchronize dda_clock and dda_step by delaying slow steps
Slow steps might take more than a TICK_TIME to make any progress. This
makes us get our velocity wrong over the duration of the step because
we only consider the velocity at the time we decide to make the step.
Fix this by delaying the step kick-off until we know exactly when it
should happen.