accel_integral: Very clean now, but no lookahead
One possible issue is that we calculate speed (dda->c) using the average
of the beginning and end velocities, but then we also slide the speed
smoothly between curr_c and next_c. Seems ok, though.

Next plan is to calculate the allowable crossover speed and store that
end_velocity in the dda.  During rampup we can calculate the time needed
to ramp down to that end_velocity when we pass it on the way up.  But if
the previous DDA was also joined, we might not pass the start velocity
again so we won't get to collect it.

So, if the start_velocity > end_velocity we can calculate the end_velocity
steps manually or with calculate_rampup_steps(), or whatever lookahead
already uses.

A more desperate option is to use all the decel-to-zero steps as our
deceleration start point, but then when we reach end_velocity we can
just continue to burn the remaining steps at that speed.  At least it will
prevent us coming to a complete stop which is quite disruptive.

Something to worry about in any case is that the "steps needed to reach
the current speed" may be calculated on one axis but then we need to
know them on some other axis because fast-axis changed when joining
moves.  In this case we might want to loop through all the calculated steps
to find the current velocity on the fast-axis so we can calculate the
steps needed. It should be cheap to do involving only addition and
uint32_t comparisons.  But there may be a faster way to find this
mathematically.