Remove ramp steps from homing movements
Homing movements schedule a huge movement which is dynamically
shortened when the endstop is hit.  It relies on rampup/rampdown
calculations which no longer affect our accel/decel phases. Also the
movement planner queue is usually several steps ahead of "real time"
so an instantaneous triggered change would require some real cleanup.

Change the homing triggers to use the new reality to detect our accel
phase. Then, if accelerating adjust total_steps so deceleration occurs
on the next planning step; if currently executing the "cruise" phase
(dda->dc=0) then short-circuit it to 1 step.  If we have decel in the
queue but we are not cruising yet, catch the cruise phase and shorten
it to one step when it is encountered in real time.