Put the DDA-being-planned responsibility in dda_queue
The dda_queue code can better keep track of which DDA we should be
planning next and which DDA is obsolete when we reset, for example.
Add a queue-following variable `mb_plan` to track the location of the
planner and let dda_queue tell the planner which DDA to plan rather
than relying on the planner itself to do it.

Also adds a dda_idle sensor so the dda_step clock can go know if it is
idle separate from the queue handler. This lets us go ahead and set
the timer for the next planned step even if the current DDA is done.
It also lets us safely try to restart the DDA timer from any point since
`dda_kick` does nothing if a timed step is already pending.