queue_step: Ignore unplanned moves
If a new move is noticed in queue_step before it is picked up in
queue_plan we will try to start it without first planning it.  This
is useless in new code which depends on the planner.

If the new move is unplanned and also the next thing we expect to handle
in queue_plan, just ignore it in step interrupt. The DDA step engine will
go idle and we will pick the movement up again in queue_plan.