Run the new move planner even if idle
When we are idle and a new move appears we init the queue with the
new move and run the planner. But we forgot to run the planner when
the queue is not idle. If the moves are very quick, we might not run
the planner in time before another move appears, and we could skip
the intervening move.