git-step-rebase: enable the rerere-mechanism.
Git provides a mechanism to record conflict solutions done with
one rebase and replaying that when rebasing another branch. That's
what's Rerere is about. Unfortunately, a rebase still stops after
resolving all conflicts with recordings, so convenience is a bit
limited.

If Rerere is disabled in a repo or there are no recorded solutions,
this option has no effect.