Simulator: be more clear about files opened.
Before, a seemingly valid command like ...

  ./sim -o sim-log.log -t 0 testcases/triangle-odd.gcode

... would not act as expected (treat both files as G-code source)
and not give meaningful hints on what's going on either. It just
reported 'Could not stat file'. No hint that it tries to open the
intended log file as G-code source.

Now the message reporting opening of a G-code file names it as
'G-code source' and opening of the recorder file gets reported,
too. Which should give a good idea about what's going on.